+++ CHANGES.md
@@ -1,5 +1,10 @@
 # Release History
 
+## 0.3.0 -- UNRELEASED
+
+Changed the logic so that completed jobs *do not* bubble up.
+Completed jobs are culled only when they are above all completed jobs.
+
 ## 0.2.0 -- 13 Oct 2017
 
 Added `Pulsar.prefix/2`: set a prefix displayed immediately before the job message.